Update setup_database.sql bigint>INTEGER

+++ setup_database.sql
@@ -4,7 +4,7 @@ CREATE TABLE users (
     username VARCHAR(255) NOT NULL UNIQUE,
     email VARCHAR(255) NOT NULL UNIQUE,
     password VARCHAR(255) NOT NULL,
-    elo BIGINT,
+    elo INTEGER,
     created_at DATE,
     updated_at DATE
 );
@@ -18,8 +18,8 @@ CREATE TABLE roles (
 -- Création de la table "users_roles" (relation users <> roles)
 CREATE TABLE users_roles (
     id BIGSERIAL PRIMARY KEY,
-    id_user BIGINT NOT NULL,
-    id_role BIGINT NOT NULL,
+    id_user INTEGER NOT NULL,
+    id_role INTEGER NOT NULL,
     FOREIGN KEY (id_user) REFERENCES users(id) ON DELETE CASCADE,
     FOREIGN KEY (id_role) REFERENCES roles(id) ON DELETE CASCADE
 );
@@ -33,13 +33,13 @@ CREATE TABLE statuses (
 -- Création de la table "matches"
 CREATE TABLE matches (
     id BIGSERIAL PRIMARY KEY,
-    player1 BIGINT NOT NULL,
-    player2 BIGINT NOT NULL,
-    id_status BIGINT,
+    player1 INTEGER NOT NULL,
+    player2 INTEGER NOT NULL,
+    id_status INTEGER,
     created_at DATE,
     updated_at DATE,
-    result_player1 BIGINT,
-    result_player2 BIGINT,
+    result_player1 INTEGER,
+    result_player2 INTEGER,
     FOREIGN KEY (player1) REFERENCES users(id) ON DELETE CASCADE,
     FOREIGN KEY (player2) REFERENCES users(id) ON DELETE CASCADE,
     FOREIGN KEY (id_status) REFERENCES statuses(id)
